remove more stuff & move out loading scene
