--- packages/client/src/scenes/loading/LoadingSceneData.ts
@@ -1,5 +1,0 @@
-import type { Connection } from "@smiley-face-game/api";
-
-export interface LoadingSceneData {
-  readonly connection: Connection;
-}
